requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-secret-manager==2.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
gunicorn==20.0.4
idna==2.10
libcst==0.3.14
mypy-extensions==0.4.3
mysqlclient==2.0.1
oauthlib==3.1.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pep8==1.7.1
Pillow==7.2.0
proto-plus==1.11.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
reportlab==3.5.59